requirements_test.txt
playwright==1.52.0
twocaptcha-python==1.1.0 